begin server refactor
